Added 6 new improvements as per Assignment # 1 (#2)
* Added a book, Corrected contribution link

1.	Correcting a broken link
Under the contributions sector the link was pointing to the wrong item. Corrected it to “CONTRIBUTIND.md”

2.	Added a new book under the general section
Added ‘Machine Learning for Text’ which promises to provide a holistic view on the topic. It was recently released and might gain popularity in the near future.

* Added a short description to Apache OpenNLP

The Apache OpenNLP API did not yet have a description. The library is great as a toolkit for text-based NLP.

* Added a new online course

Added ‘Udacity: Natural Language Processing Nanodegree’ which consists of a 3-month term encompassing the 3 projects (Part-of-Speech Tagging, Machine Translation, and a Speech recognizer)

* Added a resource under Text Summarization

Added ‘Unsupervised Text Summarization using Sentence Embeddings’ which provides a step-by-step implementation of text summarization in python.

* Added a new article under Deep Learning

Added ‘A Survey of the Usages of Deep Learning in Natural Language Processing’ which provides an overview of different DL techniques relevant for NLP.