Update README.md
Contributor: Karandeep Singh; https://github.com/Singh-K01
Below is a summary of the changes proposed to the text_mining_resources Resource:
1. Added another entry in [Blog articles, Papers, Case Studies] > Machine Translation
- [Paper Dissected: “Attention is All You Need” Explained](http://mlexplained.com/2017/12/29/attention-is-all-you-need-explained/) An important paper that first introduced attention mechanism in 2017.
This is the most important paper in Machine Translation. Presented in 2017, this introduced "attention mechanism", which was an improvement over the already cumbersome and large RNNs at the time. The architecture that this paper proposed, was computationally efficient compared to other state-of-the art architectures for Machine Translation.

2. Added another entry in [Blog articles, Papers, Case Studies] > Machine Translation
- [BERT: Pre-training of Deep Bidirectional Transformers for Language Understanding](https://arxiv.org/abs/1810.04805) A new language representation model published in 2018.
This is a new language representation model called BERT, which stands for Bidirectional Encoder Representations from Transformers. This architecture jointly conditions on both left and right context in all layers, and hence can help create state-of-the-art models for a wide range of tasks, such as question answering and language inference.

3. Added another entry in [Datasets] > Datasets
- [hate-speech-and-offensive-language](https://github.com/t-davidson/hate-speech-and-offensive-language)
This dataset has Tweets classified as hate speech, offensive language, or neither. Hate speech detection is an important consideration in social media nowadays. If you are using this data for academic research then please cite:
Thomas Davidson, Dana Warmsley, Michael Macy, and Ingmar Weber. 2017. "Automated Hate Speech Detection and the Problem of Offensive Language." Proceedings of the 11th International Conference on Web and Social Media (ICWSM). The paper is available at https://aaai.org/ocs/index.php/ICWSM/ICWSM17/paper/view/15665.

4. Added another entry in [Datasets] > Datasets
- [First Quora Dataset Release: Question Pairs](https://data.world/socialmediadata/quora-question-pairs)
The dataset that was released January 24,2017 and will give anyone the opportunity to train and test models of semantic equivalence, based on actual Quora data. 

5. Added another entry in [Blog articles, Papers, Case Studies] > Word and Document Embeddings
- [Contextual String Embeddings for Sequence Labeling](http://alanakbik.github.io/papers/coling2018.pdf). 2018.
In this paper, the authors propose to leverage the internal states of a trained character language model to produce a novel type of
word embedding which we refer to as contextual string embeddings.