Update README.md
An introduction to text analysis with Python, Part 1 – April 2012. A beginner’s walkthrough on the basics idea of sentiment analysis in Python.

A survey on sentiment analysis challenges. April 2016. A discussion and comparison of sentiment analysis challenges among forty-seven papers. 

NLP Techniques for Extracting Information. In-depth exploration of the seven steps framework of NLP data mining tools and techniques.

R or Python on Text Mining Aug 2015. Comparison of efficiency between R and Python in the field of Text Mining.

Duplicate link removal:
Cleaning
How to solve 90% of NLP problems: a step-by-step guide
Document Classification
How to solve 90% of NLP problems: a step-by-step guide (removed)